Build: Add system to copy glsl shaders to a place the executables can reach
We obviously have our shaders in the source tree. But when the user builds their demo these shaders need to be local to the executable. So we add a system to the add_concorde_executable cmake script, this will cache all of the shaders and then make a custom command for each one